Bump sqlalchemy from 1.4.46 to 2.0.0

+++ tests/database/test_database_access.py
@@ -41,7 +41,10 @@ class TestDatabaseAccess(PyironTestCase):
         Returns:
         """
         cls.database.conn.close()
-        os.remove("test_database.db")
+        if os.name != 'nt':
+            # On windows we get PermissionError: [WinError 32] The process cannot access the
+            # file because it is being used by another process: 'test_database.db'
+            os.remove("test_database.db")
 
     def tearDown(self):
         """
